test client

+++ chat-serv/serv/config.h
@@ -6,19 +6,19 @@ namespace server
 struct IServerConfig
 {
 	virtual ~IServerConfig() = default;
-	virtual size_t listen_port() const = 0;
+	virtual unsigned short listen_port() const = 0;
 };
 
 class ServerConfig : public IServerConfig
 {
 public:
 	ServerConfig();
-	virtual size_t listen_port() const override { return list_port; }
+	virtual unsigned short listen_port() const override { return list_port; }
 
 private:
 	void read_config();
 private:
-	size_t list_port;
+	unsigned short list_port;
 };
 
 }
